Formerly ../posix/gnu/types.h.~7~

+++ posix/gnu/types.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 1991 Free Software Foundation, Inc.
+/* Copyright (C) 1991, 1992 Free Software Foundation, Inc.
 This file is part of the GNU C Library.
 
 The GNU C Library is free software; you can redistribute it and/or
@@ -29,11 +29,11 @@ typedef unsigned long __u_long;
 typedef struct { long val[2]; } __quad;
 typedef struct { __u_long val[2]; } __u_quad;
 
-typedef short int __dev_t;		/* Type of device numbers.  */
+typedef int __dev_t;			/* Type of device numbers.  */
 typedef unsigned int __uid_t;		/* Type of user identifications.  */
 typedef unsigned int __gid_t;		/* Type of group identifications.  */
 typedef unsigned long int __ino_t;	/* Type of file serial numbers.  */
-typedef unsigned short int __mode_t;	/* Type of file attribute bitmasks.  */
+typedef unsigned int __mode_t;		/* Type of file attribute bitmasks.  */
 typedef unsigned short int __nlink_t;	/* Type of file link counts.  */
 typedef long int __off_t;		/* Type of file sizes and offsets.  */
 typedef int __pid_t;			/* Type of process identifications.  */
